+++ CRM.Application/Services/AuthServices/SignOutService.cs
@@ -1,4 +1,6 @@
 ﻿using CRM.Core.Entities;
+using CRM.Core.Enums;
+using CRM.Core.Exceptions;
 using CRM.Core.Interfaces.AuthServices;
 using CRM.Core.Interfaces.JwtToken;
 using CRM.Core.Interfaces.Repositories;
@@ -30,8 +32,12 @@ namespace CRM.Application.Services.AuthServices
     public async Task RemoveToken(string email, string refreshToken)
     {
       var user = await _userRepository.FindSingleAsync(e => e.Email == email);
+      if (user == null)
+        throw new CustomException(ErrorTypes.ServerError, "Server error");
 
       var removeToken = await _refreshTokenRepository.FindSingleAsync(e => e.Id == user.Id);
+      if (removeToken == null)
+        throw new CustomException(ErrorTypes.ServerError, "Server error");
 
       await _refreshTokenRepository.RemoveAsync(removeToken);
     }
